Rewrite function for finding matching pages
getData was the function that looked up a page in a wiki and, if it
found a page with a matching fullname, appended it to the interwiki. I
have renamed it to addTranslationForBranch. getDataAll is now
addTranslations for consistency.

The page category passed to interwikiFrame.html via ListPages on the
sidebar is no longer immediately stripped if it matches the branch's
static category - it is retained, and removed only when the page name
without the category is needed.

I believe my procedure for checking if a matching page exists on another
wiki despite category differences is sound, but it needs testing.
Crucially, the category of the current page will only be stripped if it
matches the current branch's category (i.e. "wanderers:" on an SCP site
without a dedicated WL wiki), and a category for the search will only be
prepended if there is a static category defined in the search candidate
branch's config.

This means for example that page on a dedicated WL wiki that are in a
non-default category will never match any page on a WL branch that does
not have a dedicated wiki - will need to make sure whether this is
desirable.

One issue that still needs solving is how to handle the wanderers-adult
category, if any such category still exists.