Switch all WL URLs to HTTPS
Wikidot usually freely interconverts HTTP/HTTPS connections, so HTTPS is
a sensible default.

Note that this should be reevaluated if more endpoints are added, as
some Wikidot endpoints are particular about how secure their connection
is. This is the case for ajax-module-connector.php, for example, but is
not the case for quickmodule.php, which is currently used.