Extract lang and community detection to parameters
Community detection (SCP vs WL) was previously performed by checking the
page category, assuming that the 'wanderers' category meant that the
page was WL and everything else was SCP. This was accurate for some
wikis that hosted both SCP and WL articles, but not for dedicated WL
sites.

Instead of attempting to determine the community in code, I have chosen
to leave that responsibility to the person implementing the Interwiki,
who will be much more accurate. The flag has been extracted to a URL
variable. This will also make it easier to test.

The current branch's language code has also been extracted to a URL
parameter. These two parameters will be the only data points needed to
determine which branch the current Interwiki is operating on.