Recover check for truncated fullnames
There was a 'shorten' check that happened if the target fullname was at
least 59 characters in length, which only took effect after the results
of the page listing had been returned, that confused me and felt
unnecessary, so I removed it. I thought it was supposed to be checking
for differences in fullname as a result of adding a category, and was
doing so incorrectly. I was correct in that this feature was missing,
and rightfully added it. However, I was wrong in what the 'shorten' flag
was actually checking for - it was actually handling the case for when a
category had been removed. I have recovered this check.