Rewrite createResizeIframe
The procedure to resize the iframe was previously being called every
50ms, forever. This is not an ideal way to go about things.

The updated version now constructs a function to resize the iframe,
which is more consistent with its name. This function will be
constructed in the main interwiki script (as opposed to in parallel with
it), and will be passed to the translation generator. It will then be
called after adding each new translation, ensuring that it is only ever
called as needed and never superfluously.