Use site URL to construct relative theme URLs
A relative URL from the perspective of the interwikiFrame resolves to
the wdfiles domain that the iframe was uploaded to. The intended use
case is for all sites to use the same HTML file, which is uploaded to a
single site in the network, so all relative styles would resolve to that
site which is obviously not tenable. (This is also why failed calls to
urlFromTheme return the empty string rather than a domain-relative URL,
as those URLs would work on one wiki and issues would go unnoticed, but
would not work on other wikis.)

This commits changes the function to accept a siteUrl argument.