Add factory for restyle function
requestStyleChange (which is called from styleFrame and also used
internally by interwikiFrame) used to have a single argument, which was
the request to parse. However, after the previous commit it is now
required to also pass the interwiki's configured base URL to the theme
URL constructor. The styleFrame cannot possibly know this information.

Instead of defining the function globally, it is now constructed during
the interwiki initialisation by createRequestStyleChange which takes the
needed URL as an argument. This is exposed to the window, so it is
effectively a global function.