Do not catch all errors when restyling
Exactly half of restyle requests are intended to fail, so any errors
emitted from them were caught and discarded. However, this naive
approach prevents actual errors from being reported.

The error handler has been changed to rethrow if the error is not a
DOMException, which is the type of error that is expected.