Add Wikidot base styling by default
I'm not totally comfortable with hardcoding any CSS into the new
interwiki, even this, but because I can guarantee that this is going to
be needed by all sites that this interwiki intends to support I think
it's a pretty reasonable addition.

When there comes a time that a site does not need the base Wikidot
styling for whatever reason - maybe a site on Wikijump, or a site hosted
elsewhere - there are other changes that would need to be made anyway.
So it's probably fine.

But just to be safe, I added a parameter to turn the automatic styling
off, which I'm not going to document.