Resize iframe after adding new styles
There is no way to tell when CSS from a link or style element has been
applied that I could find. There is a load event, and there is a
polyfill for the load event that uses the error event on images, but
both of these events fire before the CSS has actually been applied to
the document. I brute-forced it - I added a timeout.

The resizer is already debounced, so adding frivolous resizes is not a
performance concern.