Remove custom domains
Crom indexes articles by their normalised Wikidot URL without any custom
domain, so it is necessary for the branch manifest to contain this URL.
Also, for many (all?) sites, attempting to access the site by the custom
domain results in not being able to log in. EN has resolved this by
having the custom domain redirect to the Wikidot URL, but this is not
the case for all branches.

It is still necessary to distinguish between sites that support HTTPS
and those that do not, as this difference is important to the Wikidot
lookup method (PageLookupQModule).