Configure ESLint for modules
In order to use a bundler, modules are required (for import/export
syntax). However, ES5 does not support modules - this feature was
introduced in ES2015. Our ESLint config is set to restrict syntax to
ES5, so this will result in a non-negotiable linting error.

Rather than upgrade the whole project to a later syntax, a solution is
to add ESLint support for imports/exports, and then upgrade ESLint's
ecmaVersion to 2015; then, using the es plugin, disallow all ES2015
features with the exception of modules.

https://github.com/eslint/eslint/issues/9687#issuecomment-508448526

Another solution would be to use Babel to transpile the codebase to ES5,
which would enable the usage of later syntaxes.