Retain referrer protocol
The domain of the resizer iframe must be exactly identical to the domain
of the parent. Previously this was roughly achieved by using a
protocol-agnostic URL, but this was reliant on the protocol of the
interwiki iframe, which it inherited. Now that the interwiki is hosted
on GitHub Pages, it will always be https, and therefore won't work for
http sites.

Case in point: SMLT's prototype interwiki assumes http. This works fine
on http wikis, but not https wikis. This interwiki prior to this commit
effectively assumes https, which works on https wikis, but not http
wikis.

Retaining the protocol of the document referrer should resolve this.